requirements.txt
google-genai
python-dotenv
yt-dlp
pillow